make handlers extendable, add cluster configuration
